# Handle OP_MSG requests exceeding MaxMessageSizeBytes