**WIP - do not merge**

This PR tries to fix #67. All tests are passing but there's still room for a lot of improvements, especially regarding performances. I'll try to optimize this later 

**before this commit:**

```
goos: linux
goarch: amd64
BenchmarkBulkInsert-2        	     200	   6666757 ns/op	  594034 B/op	    7960 allocs/op
BenchmarkBulkInsertOpMsg-2   	     200	   6178964 ns/op	  554147 B/op	    7053 allocs/op
```
**After this commit:**

```
goos: linux
goarch: amd64
BenchmarkBulkInsert-2        	     200	   6552978 ns/op	  663142 B/op	    8961 allocs/op
BenchmarkBulkInsertOpMsg-2   	     200	   6405799 ns/op	  684696 B/op	    8066 allocs/op
```
using this code for benchmarking: 

```
	s, err := mgo.Dial("mongodb://localhost:27017")
	if err != nil {
		b.Fail()
	}
	c := s.DB("bench").C("test")
	b.ResetTimer()

	for n := 0; n < b.N; n++ {
		bulk := c.Bulk()
		bulk.Unordered()
		for i := 0; i < 1000; i++ {
			bulk.Insert(bson.M{"n": i})
		}
		_, err := bulk.Run()
		if err != nil {
			b.Fail()
		}
	}
```

Any comments / review on this is welcome!