@domodwyer 2017-12-20
Hi @feliixx 

Sorry for the delay - I'm away without internet till the new year, I'll be able to do a proper review when I'm back with a laptop but looks good so far.

Thanks for the benchmarks!

Dom