@domodwyer 2017-12-27
Hi @feliixx 

Sorry for the delay, back in the internet connected world again!

I've been using this branch on a side project as a test and it works great, thanks for the PR. My only concern against merging is the error handling - previously marshalling the BSON was handled when the user called `Run()` but this PR shifts the error for `Marshal()` to the individual operation methods (`Bulk.Insert()` / `Bulk.Update()` / etc). I understand they're unchecked because the methods do not return an error, and need to be backwards compatible - what do you think to keeping the marshalling as you have it now, but somehow adding errors during marshalling to the `BulkError` returned by `Run()`? As it stands, a value that cannot be marshalled is silently dropped from the bulk request.

The existing behaviour is tricky/inconsistent though - if the operation is unordered, the other operations continue and the error is returned at the end - this is the easy case! If the bulk operation is ordered, when the first operation fails the bulk request is stopped and the error is returned immediately - even though we know an error is "going to happen" we should maintain this behaviour as some may rely on it - what do you think?

I'm happy to help with the above, just wanted your opinion!

Dom