@feliixx 2017-12-28
my initial idea was that `marshal()` always return an (non-empty) array of bytes, even if the document can't be marshalled. So the corrupted bytes are send to the database anyway, and then MongoDB returns an error when trying to unmarshall the message. 
So basically, instead of checking if the document is correct before sending it, we send the document and let the database deal with the error checking

Because all the tests are passing I though it was working fine, I'll add a new test to make sure it works as intended