@domodwyer 2018-01-08
Hi @feliixx!

Would you mind retargeting this to `feature/OP_MSG-support` due to #73?

I'm going to be profiling the OP_MSG support branch and trying to resolve any issues hopefully in the next couple weeks, I'll happily work with you if you have any ideas though?

Dom